kafka admin and producer config

+++ clients/src/main/java/com/gustyflows/clients/fraud/FraudServiceClient.java
@@ -12,5 +12,5 @@ public interface FraudServiceClient {
     //this is our fraud-service client, and an interface which will target FraudController
     //any microservice that wants to talk to fraud-service, can use this interface
     @GetMapping(path = "api/v1/fraud-check/{customerId}")
-    FraudCheckResponse isFraudster(@PathVariable("customerId") Integer customerID);
+    FraudCheckResponse isFraudster(@PathVariable("customerId") Long customerID);
 }
